+++ src/Sulu/Bundle/AdminBundle/Resources/public/js/components/app/main.js
@@ -150,14 +150,14 @@ define(function() {
          * @param {object} request
          * @return {string}
          */
-        extractErrorMessage: function (request) {
+        extractErrorMessage: function(request) {
             var message = [request.status];
 
             // if response is symfony JSON exception
             if (request.responseJSON !== undefined) {
                 var response = request.responseJSON
 
-                this.sandbox.util.each(response, function (index) {
+                this.sandbox.util.each(response, function(index) {
                     var exception = response[index];
 
                     if (exception.message !== undefined) {
@@ -221,17 +221,17 @@ define(function() {
                 // prevent the default action for the anchor tag
                 this.sandbox.dom.preventDefault(event);
 
-                var dataSuluEvent = this.sandbox.dom.attr(event.currentTarget, 'data-sulu-event');
+                var dataSuluEvent = this.sandbox.dom.attr(event.currentTarget, 'data-sulu-event'),
+                    eventArgs = this.sandbox.dom.data(event.currentTarget, 'eventArgs');
 
                 // if data-sulu-event attribute is set emit the attribute value as an event
                 if (!!dataSuluEvent &&
                     typeof dataSuluEvent === 'string') {
-                    this.sandbox.emit(dataSuluEvent);
+                    this.sandbox.emit(dataSuluEvent, eventArgs);
                 }
 
                 // if valid href attribute is set navigate to it using the sulu.navigate method
-                if (!!event.currentTarget.attributes.href &&
-                    !!event.currentTarget.attributes.href.value &&
+                if (!!event.currentTarget.attributes.href && !!event.currentTarget.attributes.href.value &&
                     event.currentTarget.attributes.href.value !== '#') {
 
                     this.emitNavigationEvent({action: event.currentTarget.attributes.href.value}, true, true);
@@ -302,6 +302,13 @@ define(function() {
                     this.setTitlePostfix(this.sandbox.translate(event.title));
                 }
             }.bind(this));
+
+            // FIXME aura cannot handle so fast navigates (https://github.com/sulu-cmf/sulu/issues/892)
+            this.sandbox.on('husky.data-navigation.select', _.debounce(function(item) {
+                if (!!item && !!item._links && !!item._links.admin) {
+                    this.sandbox.emit('sulu.router.navigate', item._links.admin.href, true, false);
+                }
+            }.bind(this), 1000));
 
             // content tabs event
             this.sandbox.on('husky.tabs.content.item.select', function(event) {
@@ -493,7 +500,7 @@ define(function() {
             //Todo: don't use hardcoded url
             this.navigate('contacts/contacts/edit:' + this.sandbox.sulu.user.contact.id + '/details', true, false, false);
             this.sandbox.emit('husky.navigation.select-item', 'contacts/contacts');
-         },
+        },
 
         /**
          * Takes a postifix and updates the page title