+++ src/Sulu/Bundle/MediaBundle/Admin/SuluMediaAdmin.php
@@ -11,6 +11,7 @@
 namespace Sulu\Bundle\MediaBundle\Admin;
 
 use Sulu\Bundle\AdminBundle\Admin\Admin;
+use Sulu\Bundle\AdminBundle\Navigation\DataNavigationItem;
 use Sulu\Bundle\AdminBundle\Navigation\Navigation;
 use Sulu\Bundle\AdminBundle\Navigation\NavigationItem;
 use Sulu\Component\Security\Authorization\SecurityCheckerInterface;
@@ -33,8 +34,16 @@ class SuluMediaAdmin extends Admin
         $media->setIcon('image');
 
         if ($this->securityChecker->hasPermission('sulu.media.collections', 'view')) {
-            $collections = new NavigationItem('navigation.media.collections', $media);
-            $collections->setAction('media/collections');
+            $collections = new DataNavigationItem('navigation.media.collections', '/admin/api/collections', $media);
+            $collections->setId('collections-edit');
+            $collections->setAction('media/collections/root');
+            $collections->setInstanceName('collections');
+            $collections->setDataNameKey('title');
+            $collections->setDataResultKey('collections');
+            $collections->setShowAddButton(true);
+            $collections->setTitleTranslationKey('navigation.media.collections');
+            $collections->setNoDataTranslationKey('');
+            $collections->setAddButtonTranslationKey('navigation.media.collections.add');
         }
 
         if ($media->hasChildren()) {