+++ src/Sulu/Bundle/MediaBundle/Resources/public/js/components/media-selection/main.js
@@ -41,7 +41,8 @@ define(['sulumedia/collection/collections', 'sulumedia/model/collection'], funct
         },
 
         constants = {
-            lastVisitedCollectionKey: 'last-visited-collection'
+            lastVisitedCollectionKey: 'last-visited-collection',
+            listViewStorageKey: 'mediaOverlayListView'
         },
 
         /**
@@ -76,30 +77,19 @@ define(['sulumedia/collection/collections', 'sulumedia/model/collection'], funct
         },
 
         templates = {
-            addTab: function(options, header) {
+            mediaSelection: function(options) {
                 return [
-                    '<div id="', options.ids.chooseTab, '">',
-                    '   <div class="heading">',
-                    '       <h3>', header, '</h3>',
+                    '<div class="media-selection-overlay">',
+                    '   <div class="media-selection-overlay-navigation-container pull-left"></div>',
+                    '   <div class="media-selection-overlay-content">',
+                    '       <div class="fa-times media-selection-overlay-close"></div>',
+                    '       <div class="media-selection-overlay-dropzone-container"></div>',
+                    '       <div class="media-selection-overlay-toolbar-container"></div>',
+                    '       <div class="media-selection-overlay-content-title">' + options.contentDefaultTitle + '</div>',
+                    '       <div class="media-selection-overlay-datagrid-container"></div>',
                     '   </div>',
-                    '   <div id="', options.ids.gridGroup, '"/>',
-                    '   <div class="overlay-loader" id="', options.ids.loader , '"></div>',
                     '</div>'
                 ].join('');
-            },
-
-            uploadTab: function(options, collection) {
-                return [
-                    '<div id="', options.ids.uploadTab , '">',
-                    '   <div class="grid-row">',
-                    '       <label>', collection , '</label>',
-                    '       <div id="', options.ids.collectionSelect , '"></div>',
-                    '   </div>',
-                    '   <div class="grid-row">',
-                    '       <div id="', options.ids.dropzone , '"></div>',
-                    '   </div>',
-                    '</div>'
-                ].join('')
             },
 
             contentItem: function(title, thumbnails) {
@@ -144,6 +134,53 @@ define(['sulumedia/collection/collections', 'sulumedia/model/collection'], funct
         },
 
         /**
+         * Scrolls the whole form the the bottom
+         */
+        scrollToBottom = function() {
+            this.sandbox.dom.scrollAnimate(this.sandbox.dom.height('.media-selection-overlay-datagrid-container'), '.media-selection-overlay-content');
+        },
+
+        /**
+         * Takes an array of files and adds them to the datagrid
+         * @param files {Array} array of files
+         */
+        addFilesToDatagrid = function(files) {
+            for (var i = -1, length = files.length; ++i < length;) {
+                files[i].selected = true;
+            }
+            this.sandbox.emit('husky.datagrid.media-selection-ovelay.' + this.options.instanceName + '.records.add', files, scrollToBottom.bind(this));
+            this.sandbox.emit('husky.data-navigation.' + this.options.instanceName + '.collections.reload');
+        },
+
+        /**
+         * Handler method for the data navigation select event
+         * @method dataNavigationSelectHandler
+         * @param {Object} collection
+         */
+        dataNavigationSelectHandler = function(collection) {
+            var collectionId, 
+                collectionTitle = this.sandbox.translate('media-selection.overlay.all-images');
+
+            if (collection) {
+                collectionId = collection.id;
+                collectionTitle = collection.title;
+
+                this.sandbox.emit('husky.toolbar.media-selection-ovelay.' + this.options.instanceName + '.item.enable', 'add');
+                this.sandbox.emit('husky.dropzone.media-selection-ovelay.' + this.options.instanceName + '.enable');
+            } else {
+                this.sandbox.emit('husky.toolbar.media-selection-ovelay.' + this.options.instanceName + '.item.disable', 'add');
+                this.sandbox.emit('husky.dropzone.media-selection-ovelay.' + this.options.instanceName + '.disable');
+            }
+
+            this.sandbox.emit('husky.datagrid.media-selection-ovelay.' + this.options.instanceName + '.url.update', { 
+                collection: collectionId,
+                page: 1
+            });
+            changeUploadCollection.call(this, collectionId);
+            this.$el.find('.media-selection-overlay-content-title').html(collectionTitle);
+        },
+
+        /**
          * custom event handling
          */
         bindCustomEvents = function() {
@@ -160,20 +197,19 @@ define(['sulumedia/collection/collections', 'sulumedia/model/collection'], funct
                 setData.call(this, {ids: ids}, false);
             }, this);
 
-            this.sandbox.on('husky.tabs.overlaymedia-selection.' + this.options.instanceName + '.add.initialized', function() {
+            this.sandbox.on('husky.tabs.overlay.' + this.options.instanceName + '.add.initialized', function() {
                 startOverlayLoader.call(this);
                 this.collections.fetchSorted('title', {
                     success: function(collections) {
                         this.collectionArray = collections.toJSON();
                         stopOverlayLoader.call(this);
                         startGridGroup.call(this);
-                        startSelect.call(this);
                         startDropzone.call(this);
                     }.bind(this)
                 });
             }.bind(this));
 
-            this.sandbox.on('husky.overlay.media-selection.' + this.options.instanceName + '.add.opened', function() {
+            this.sandbox.on('husky.overlay.' + this.options.instanceName + '.add.opened', function() {
                 if (this.gridGroupDeprecated === true) {
                     reloadGridGroup.call(this);
                     this.gridGroupDeprecated = false;
@@ -182,19 +218,19 @@ define(['sulumedia/collection/collections', 'sulumedia/model/collection'], funct
 
             // set position of overlay if height of grid-group changes
             this.sandbox.on('sulu.grid-group.' + this.options.instanceName + '.height-changed', function() {
-                this.sandbox.emit('husky.overlay.media-selection.' + this.options.instanceName + '.add' + '.set-position');
+                this.sandbox.emit('husky.overlay.' + this.options.instanceName + '.add' + '.set-position');
             }.bind(this));
 
             // set position of overlay if grid-group has initialized
             this.sandbox.on('sulu.grid-group.' + this.options.instanceName + '.initialized', function() {
-                this.sandbox.emit('husky.overlay.media-selection.' + this.options.instanceName + '.add' + '.set-position');
+                this.sandbox.emit('husky.overlay.' + this.options.instanceName + '.add' + '.set-position');
             }.bind(this));
 
             // save the collection where new media should be uploaded
-            this.sandbox.on('husky.select.media-selection-' + this.options.instanceName + '.selected.item', changeUploadCollection.bind(this));
+            this.sandbox.on('husky.select.' + this.options.instanceName + '.selected.item', changeUploadCollection.bind(this));
 
             // add uploaded files
-            this.sandbox.on('husky.dropzone.media-selection-' + this.options.instanceName + '.files-added', addUploadedFile.bind(this));
+            // this.sandbox.on('husky.dropzone.' + this.options.instanceName + '.files-added', addUploadedFile.bind(this));
 
             this.sandbox.on('sulu.grid-group.' + this.options.instanceName + '.record-selected', function(id) {
                 this.sandbox.emit(RECORD_SELECTED.call(this), id);
@@ -202,6 +238,82 @@ define(['sulumedia/collection/collections', 'sulumedia/model/collection'], funct
 
             this.sandbox.on('sulu.grid-group.' + this.options.instanceName + '.record-deselected', function(id) {
                 this.sandbox.emit(RECORD_DESELECTED.call(this), id);
+            }.bind(this));
+
+            this.sandbox.on('husky.data-navigation.' + this.options.instanceName + '.select', dataNavigationSelectHandler.bind(this));
+
+            // change datagrid to table
+            this.sandbox.on('sulu.list-toolbar.media-selection-ovelay.' + this.options.instanceName + '.change.table', function() {
+                this.sandbox.emit('husky.datagrid.media-selection-ovelay.' + this.options.instanceName + '.view.change', 'table');
+                this.sandbox.sulu.saveUserSetting(constants.listViewStorageKey, 'table');
+            }.bind(this));
+
+            // change datagrid to thumbnail small
+            this.sandbox.on('sulu.list-toolbar.media-selection-ovelay.' + this.options.instanceName + '.change.thumbnail-small', function() {
+                this.sandbox.emit('husky.datagrid.media-selection-ovelay.' + this.options.instanceName + '.view.change', 'thumbnail', {large: false});
+                this.sandbox.sulu.saveUserSetting(constants.listViewStorageKey, 'thumbnailSmall');
+            }.bind(this));
+
+            // change datagrid to thumbnail large
+            this.sandbox.on('sulu.list-toolbar.media-selection-ovelay.' + this.options.instanceName + '.change.thumbnail-large', function() {
+                this.sandbox.emit('husky.datagrid.media-selection-ovelay.' + this.options.instanceName + '.view.change', 'thumbnail', {large: true});
+                this.sandbox.sulu.saveUserSetting(constants.listViewStorageKey, 'thumbnailLarge');
+            }.bind(this));
+
+            // load collections list if back icon is clicked
+            this.sandbox.on('sulu.header.back', function() {
+                this.sandbox.emit('sulu.media.collections.list');
+            }.bind(this));
+
+            // if files got uploaded to the server add them to the datagrid
+            this.sandbox.on('husky.dropzone.media-selection-ovelay.' + this.options.instanceName + '.files-added', function(files) {
+                this.sandbox.emit('sulu.labels.success.show', 'labels.success.media-upload-desc', 'labels.success');
+                addFilesToDatagrid.call(this, files);
+            }.bind(this));
+
+            // open data-source folder-overlay
+            this.sandbox.on('sulu.list-toolbar.media-selection-ovelay.' + this.options.instanceName + '.add', function() {
+                this.sandbox.emit('husky.dropzone.media-selection-ovelay.' + this.options.instanceName + '.open-data-source');
+            }.bind(this));
+
+            // add image to the selected images grid
+            this.sandbox.on('husky.datagrid.media-selection-ovelay.' + this.options.instanceName + '.item.select', function(itemId) {
+                var data = this.getData(),
+                    index = data.ids.indexOf(itemId);
+
+                if (index > -1) {
+                    return;
+                }
+
+                data.ids.push(itemId);
+                this.setData(data);
+                reloadGridGroup.call(this)
+            }.bind(this));
+
+            // remove image to the selected images grid
+            this.sandbox.on('husky.datagrid.media-selection-ovelay.' + this.options.instanceName + '.item.deselect', function(itemId) {
+                var data = this.getData(),
+                    index = data.ids.indexOf(itemId);
+
+                if (index > -1) {
+                    data.ids.splice(index, 1);
+                }
+
+                this.setData(data);
+                reloadGridGroup.call(this)
+            }.bind(this));
+
+            this.sandbox.on('husky.overlay.dropzone-media-selection-ovelay.' + this.options.instanceName + '.opened', function() {
+                this.$el.find('.media-selection-overlay-container').addClass('dropzone-overlay-opened');
+            }.bind(this));
+
+            this.sandbox.on('husky.overlay.dropzone-media-selection-ovelay.' + this.options.instanceName + '.closed', function() {
+                this.$el.find('.media-selection-overlay-container').removeClass('dropzone-overlay-opened');
+            }.bind(this));
+
+            this.sandbox.on('husky.overlay.' + this.options.instanceName + '.add.opened', function() {
+                var selectedItems = this.getData().ids || [];
+                this.sandbox.emit('husky.datagrid.media-selection-ovelay.' + this.options.instanceName + '.selected.update', selectedItems);
             }.bind(this));
         },
 
@@ -212,7 +324,7 @@ define(['sulumedia/collection/collections', 'sulumedia/model/collection'], funct
         changeUploadCollection = function(collectionId) {
             this.uploadCollection = collectionId;
             this.sandbox.emit(
-                    'husky.dropzone.media-selection-' + this.options.instanceName + '.change-url',
+                    'husky.dropzone.media-selection-ovelay.' + this.options.instanceName + '.change-url',
                     '/admin/api/media?collection=' + collectionId);
         },
 
@@ -282,31 +394,6 @@ define(['sulumedia/collection/collections', 'sulumedia/model/collection'], funct
         },
 
         /**
-         * Starts the select for choosing the collection to upload
-         */
-        startSelect = function() {
-            var data = this.sandbox.util.extend([], true, this.collectionArray),
-                preselected = this.sandbox.sulu.getUserSetting(constants.lastVisitedCollectionKey) || 'new';
-            data.unshift({
-                id: 'new',
-                title: this.sandbox.translate(this.options.translations.createNewCollection)
-            });
-            changeUploadCollection.call(this, preselected);
-            this.sandbox.start([
-                {
-                    name: 'select@husky',
-                    options: {
-                        el: getId.call(this, 'collectionSelect'),
-                        instanceName: 'media-selection-' + this.options.instanceName,
-                        valueName: 'title',
-                        data: data,
-                        preSelectedElements: [preselected]
-                    }
-                }
-            ]);
-        },
-
-        /**
          * Starts the dropzone for uploading a new media
          */
         startDropzone = function() {
@@ -319,7 +406,7 @@ define(['sulumedia/collection/collections', 'sulumedia/model/collection'], funct
                         method: 'POST',
                         paramName: 'fileVersion',
                         showOverlay: false,
-                        instanceName: 'media-selection-' + this.options.instanceName,
+                        instanceName: 'media-selection-ovelay.' + this.options.instanceName,
                         afterDropCallback: uploadNewFile.bind(this),
                         keepFilesAfterSuccess: true
                     }
@@ -395,7 +482,7 @@ define(['sulumedia/collection/collections', 'sulumedia/model/collection'], funct
                 }.bind(this));
                 this.setData(data);
                 this.sandbox.emit('sulu.labels.success.show', 'labels.success.media-upload-desc', 'labels.success');
-                reloadGridGroup.call(this)
+                reloadGridGroup.call(this);
             }
         },
 
@@ -403,9 +490,6 @@ define(['sulumedia/collection/collections', 'sulumedia/model/collection'], funct
          * starts the overlay component
          */
         startAddOverlay = function() {
-            var chooseTabData = templates.addTab(this.options, this.sandbox.translate(this.options.translations.collections)),
-                uploadTabData = templates.uploadTab(this.options, this.sandbox.translate(this.options.translations.collection));
-
             var $element = this.sandbox.dom.createElement('<div/>');
             this.sandbox.dom.append(this.$el, $element);
 
@@ -414,33 +498,187 @@ define(['sulumedia/collection/collections', 'sulumedia/model/collection'], funct
                     name: 'overlay@husky',
                     options: {
                         triggerEl: this.$addButton,
-                        cssClass: 'media-selection-overlay',
-                        el: $element,
-                        removeOnClose: false,
-                        container: this.$el,
                         draggable: false,
-                        instanceName: 'media-selection.' + this.options.instanceName + '.add',
-                        skin: 'medium',
+                        removeOnClose: false,
+                        el: $element,
+                        container: this.$el,
+                        cssClass: 'media-selection-overlay-container',
+                        instanceName: this.options.instanceName + '.add',
+                        skin: 'wide',
+                        supportKeyInput: false,
                         slides: [
                             {
                                 title: this.sandbox.translate(this.options.translations.addImages),
-                                okCallback: getAddOverlayData.bind(this),
-                                cssClass: 'media-selection-overlay-add',
-                                tabs: [
-                                    {
-                                        title: this.sandbox.translate(this.options.translations.choose),
-                                        data: chooseTabData
-                                    },
-                                    {
-                                        title: this.sandbox.translate(this.options.translations.upload),
-                                        data: uploadTabData
-                                    }
-                                ]
+                                data: templates.mediaSelection({
+                                    contentDefaultTitle: this.sandbox.translate('media-selection.overlay.all-images')
+                                })
                             }
                         ]
                     }
                 }
             ]);
+
+            this.sandbox.once('husky.overlay.' + this.options.instanceName + '.add.opened', function() {
+                
+                this.$el.on('click', '.media-selection-overlay-close', function() {
+                    this.sandbox.emit('husky.overlay.' + this.options.instanceName + '.add.close');
+                }.bind(this));
+
+                this.sandbox.start([
+                    {
+                        name: 'data-navigation@husky',
+                        options: {
+                            el: this.$el.find('.media-selection-overlay-navigation-container'),
+                            resultKey: 'collections',
+                            showAddButton: false,
+                            rootUrl: '/admin/api/collections',
+                            url: '/admin/api/collections',
+                            nameKey: 'title',
+                            instanceName: this.options.instanceName,
+                            globalEvents: false
+                        }
+                    }
+                ]);
+
+                this.sandbox.sulu.initListToolbarAndList.call(
+                    this, 
+                    'mediaOverlay', 
+                    [
+                        {
+                            name: 'id',
+                            translation: 'public.id',
+                            disabled: true,
+                            default: false,
+                            sortable: true,
+                            type: '',
+                            width: '50px',
+                            minWidth: '',
+                            editable: false,
+                            class: ''
+                        },
+                        {
+                            name: 'thumbnails',
+                            translation: 'media.media.thumbnails',
+                            disabled: false,
+                            default: true,
+                            sortable: true,
+                            type: 'thumbnails',
+                            width: '',
+                            minWidth: '',
+                            editable: false,
+                            class: ''
+                        },
+                        {
+                            name: 'title',
+                            translation: 'public.title',
+                            disabled: false,
+                            default: false,
+                            sortable: true,
+                            type: 'title',
+                            width: '',
+                            minWidth: '',
+                            editable: false,
+                            class: ''
+                        },
+                        {
+                            name: 'size',
+                            translation: 'media.media.size',
+                            disabled: false,
+                            default: true,
+                            sortable: true,
+                            type: 'bytes',
+                            width: '',
+                            minWidth: '',
+                            editable: false,
+                            class: ''
+                        }
+                    ],
+                    {
+                        el: this.$el.find('.media-selection-overlay-toolbar-container'),
+                        instanceName: 'media-selection-ovelay.' + this.options.instanceName,
+                        template: [
+                            {
+                                id: 'add',
+                                icon: 'plus-circle',
+                                disabled: true,
+                                callback: function() {
+                                    this.sandbox.emit('husky.dropzone.media-selection-ovelay.' + this.options.instanceName + '.open-data-source');
+                                }.bind(this)
+                            },
+                            {
+                                id: 'change',
+                                icon: 'th-large',
+                                itemsOption: {
+                                    markable: true
+                                },
+                                items: [
+                                    {
+                                        id: 'small-thumbnails',
+                                        title: this.sandbox.translate('sulu.list-toolbar.small-thumbnails'),
+                                        callback: function() {
+                                            this.sandbox.emit('sulu.list-toolbar.media-selection-ovelay.' + this.options.instanceName + '.change.thumbnail-small');
+                                        }.bind(this)
+                                    },
+                                    {
+                                        id: 'big-thumbnails',
+                                        title: this.sandbox.translate('sulu.list-toolbar.big-thumbnails'),
+                                        callback: function() {
+                                            this.sandbox.emit('sulu.list-toolbar.media-selection-ovelay.' + this.options.instanceName + '.change.thumbnail-large');
+                                        }.bind(this)
+                                    },
+                                    {
+                                        id: 'table',
+                                        title: this.sandbox.translate('sulu.list-toolbar.table'),
+                                        callback: function() {
+                                            this.sandbox.emit('sulu.list-toolbar.media-selection-ovelay.' + this.options.instanceName + '.change.table');
+                                        }.bind(this)
+                                    }
+                                ]
+                            }
+                        ],
+                        inHeader: false
+                    },
+                    {
+                        el: this.$el.find('.media-selection-overlay-datagrid-container'),
+                        url: '/admin/api/media',
+                        view: 'thumbnail',
+                        resultKey: 'media',
+                        instanceName: 'media-selection-ovelay.' + this.options.instanceName,
+                        preselected: this.getData().ids,
+                        viewOptions: {
+                            table: {
+                                fullWidth: false,
+                                rowClickSelect: true
+                            },
+                            thumbnail: {
+                                large: false,
+                                unselectOnBackgroundClick: false
+                            }
+                        },
+                        paginationOptions: {
+                            dropdown: {
+                                verticalAlignment: 'top'
+                            }
+                        }
+                    }
+                );
+                
+                this.sandbox.start([
+                    {
+                        name: 'dropzone@husky',
+                        options: {
+                            el: this.$el.find('.media-selection-overlay-dropzone-container'),
+                            url: '/admin/api/media',
+                            method: 'POST',
+                            paramName: 'fileVersion',
+                            instanceName: 'media-selection-ovelay.' + this.options.instanceName,
+                            dropzoneEnabled: false,
+                            cancelUploadOnOverlayClick: true
+                        }
+                    }
+                ]);
+
+            }.bind(this));
         },
 
         /**