+++ src/Sulu/Bundle/SecurityBundle/Resources/public/js/components/users/components/form/main.js
@@ -421,7 +421,11 @@ define(['app-config', 'widget-groups'], function(AppConfig, WidgetGroups) {
                 }
             ]);
 
-            this.sandbox.util.load(constants.localizationsUrl, null, 'json').then(this.renderRoles.bind(this));
+            this.sandbox.util.load(constants.localizationsUrl, null, 'json')
+                .then(function(response) {
+                    return response._embedded.localizations;
+                })
+                .then(this.renderRoles.bind(this));
         },
 
         renderRoles: function (localizations) {